chore: rename header guard names to contain the timestamp not filename
This way the header guard name is independent from the filename (so it will not
have to be kept up to date with renaming). If a project has multiple authors
the timestamp might have to be combined with the author name.

The name is chosen to be likely unique and unlikely to be updated.
The github username, library name and file ending are considered likely to
change. While the project name might change, it is unlikely.

The timestamps for this commit are not the real file creation timestamps
because they are not available, but taken from version control. The minute is
incremented by one for each file in a commit with multiple file creations.