feat: init tts and SM after UI for faster startup and navigation
Creating the tts instance and the state machine made the
navigation laggy and the startup take longer.
While it was considered to use threads, the onCompleted handler
in qml needed less code.
The disadvantage with the latter are the Q_INVOKABLEs that are
available from QML to accidently get called again.