chore: use default constructor for system locale
This prepares for unit testing, since now the actual system
locale can be 'overridden' by setting the default locale.

Maybe using the default locale is not as clean as creating a
mock object? Advantage is that it takes less code.
Disadvantage might be that the default locale might also get
used in unexpected places once set. And that it can accidentally
get set in the non-test code.