chore: set qm search dir for auto locale via parameter
This prepares for unit testing. While I want to look for resource
files in the Qt Resource System to be independent from the file system
structure, it might not be available when unit testing.

Additionally, this commit removes the dependency on the resource prefix.
Though starting from the base resource path ':' could maybe lead to
performance issues if the app gets more resources?