chore: use settings member instead of inheritance
The inheritance was like a hack for settings. While the methods
should be accessible by the translator, and I intended to make
the functionality of the Translator class clearer by having a
list of classes it inherits functionality from,
this now seems hacky to me and a member variable seems clearer.