chore: add member for key in resource list
Auto locale can be resolved to system locale, which
could e.g. use a different territory (especially in case of
having a resource with AnyTerritory) than the resource entry it
resolves to.
Save the key of the corresponding resource entry in case the
resource path is needed later on for e.g. loading a
translation.